+++ src/modules/boolean.js
@@ -1,11 +1,11 @@
-import { updateIndex } from './showList.js';
+import { updateIndex } from './updateIndex.js';
 
 const trueStatus = (task) => {
-    task.completed = true;
-}
+  task.completed = true;
+};
 const falseStatus = (task) => {
-    task.completed = false;
-}
+  task.completed = false;
+};
 
 const clearCompleted = (bArray) => {
   bArray = bArray.filter((task) => !task.completed);
