+++ src/modules/updateIndex.js
@@ -1,0 +1,10 @@
+/* eslint-disable */
+let array = JSON.parse(localStorage.getItem('Data')) || [];
+
+const updateIndex = () => {
+    array.forEach((task, arrayIndex) => {
+      task.index = arrayIndex;
+    });
+};
+  
+export { updateIndex };
